rename member.dart to rider.dart and move to model/rider/rider.dart
