+++ lib/database/export_rides_dao.dart
@@ -1,9 +1,9 @@
 import 'package:sembast/sembast.dart';
 import 'package:weforza/database/database_tables.dart';
 import 'package:weforza/model/export/exportable_ride.dart';
-import 'package:weforza/model/member.dart';
 import 'package:weforza/model/ride.dart';
 import 'package:weforza/model/ride_attendee.dart';
+import 'package:weforza/model/rider/rider.dart';
 
 abstract class ExportRidesDao {
   /// Get the exportable rides with their attendees.
