+++ lib/file/import_riders_csv_reader.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:weforza/file/import_riders_file_reader.dart';
 import 'package:weforza/model/device/device.dart';
-import 'package:weforza/model/member.dart';
+import 'package:weforza/model/rider/rider.dart';
 import 'package:weforza/model/rider/serializable_rider.dart';
 
 /// This class represents an [ImportRidersFileReader]
