+++ lib/model/rider/rider_form_delegate.dart
@@ -1,6 +1,6 @@
 import 'package:uuid/uuid.dart';
 import 'package:weforza/model/async_computation_delegate.dart';
-import 'package:weforza/model/member.dart';
+import 'package:weforza/model/rider/rider.dart';
 import 'package:weforza/model/rider/rider_model.dart';
 import 'package:weforza/repository/member_repository.dart';
