+++ lib/widgets/pages/rider_form.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:weforza/exceptions/exceptions.dart';
 import 'package:weforza/generated/l10n.dart';
-import 'package:weforza/model/member.dart';
 import 'package:weforza/model/profile_image_picker_delegate.dart';
+import 'package:weforza/model/rider/rider.dart';
 import 'package:weforza/model/rider/rider_form_delegate.dart';
 import 'package:weforza/model/rider/rider_model.dart';
 import 'package:weforza/model/rider/rider_validator.dart';
